pipeline folder/requirements.txt
os
re
glob
json
boto3
python-dotenv
pandas
psycopg2-binary